requirements.txt
# ── Streamlit UI ──────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── AI / ML ───────────────────────────────────────────────────────────────────
google-genai>=1.0.0
google-generativeai>=0.7.0   # kept for backend/routers/analyze.py

# ── Data / Validation ─────────────────────────────────────────────────────────
pydantic>=2.0.0
python-dotenv>=1.0.0

# ── Backend API (kept for local dev / reference — NOT needed on Streamlit Cloud)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
httpx>=0.27.0
